Attempt to get the OpenAI Real-Time Chat Demo runnin on the Waveshare ESP32-P4 Touch LCD (Round 4" 720x720). With these changes, it builds and mostly runs.
If you are using the BSP, you get output configured as is:

        .mic_layout = "MMR",           // Mic, Mic, Reference layout for 3 channels
        .channel = 3,                  // 3 channels: 2 mics, MIC3 is reference from ES8311 loopback
        .channel_mask = 1 | 2 | 4,     // Enable channels

If you use the new board definition, you have to switch to 2 channels to be able to interact with it.

        .channel = 2,                  // 2 channels: 2 mics
        .channel_mask = 1 | 2,         // Enable channels

In either case, I can't figure out how to Acoustic Echo Cancellation (AEC) working so right now it talks to itself as the output is taken as input.
Posting so I can hopefully get some pointers on how to correctly configure it.

The changes I made were to:

Add a new board definition for the # Waveshare ESP32-P4 Touch LCD (Round 4" 720x720).
Add logic in build_capture_system() to initialize the mic layout and channels to utilize mic1 and mic2 with mic3 as the reference from ES8311
Update the TEST_BOARD_NAME to utilize the newly added WAVESHARE_ESP32P4_ROUND_LCD

Update media_sys.c to be able to use the Waveshare BSP when you #define USE_BSP_AUDIO_CODEC (otherwise it uses the new board definition).

Add a model partition so I can utilize Wake Word Detection and Voice Activity Detection